add "experimental" as a parameter

+++ download.sh
@@ -1,5 +1,14 @@
-svn co https://repository.napa-wine.eu/svn/napawine-software/trunk/Applications/OfferStreamer || exit
-svn co https://repository.napa-wine.eu/svn/napawine-software/trunk/GRAPES || exit
+if [[ $1 && $1 == "experimental" ]]; then
+  touch experimental
+fi
+
+if [ ! -f experimental ]; then
+  svn co https://repository.napa-wine.eu/svn/napawine-software/trunk/Applications/OfferStreamer || exit
+  svn co https://repository.napa-wine.eu/svn/napawine-software/trunk/GRAPES || exit
+else
+  git clone http://www.disi.unitn.it/~kiraly/SharedGits/Streamers.git || exit; mv Streamers OfferStreamer || exit
+  git clone http://www.disi.unitn.it/~kiraly/PublicGits/GRAPES.git || exit
+fi
 
 # get libevent 2.0.3 (warning: newer version were not tested)
 wget http://www.monkey.org/~provos/libevent-2.0.3-alpha.tar.gz || exit
@@ -12,7 +21,11 @@ tar xvzf confuse-2.7.tar.gz || exit
 #cd confuse; git checkout V2_7; cd ..
 
 #get ffmpeg
+if [ ! -f experimental ]; then
+  svn checkout svn://svn.ffmpeg.org/ffmpeg/trunk -r 23732 ffmpeg
+else
   (wget http://ffmpeg.org/releases/ffmpeg-checkout-snapshot.tar.bz2; tar xjf ffmpeg-checkout-snapshot.tar.bz2; mv ffmpeg-checkout-20* ffmpeg) || svn checkout svn://svn.ffmpeg.org/ffmpeg/trunk ffmpeg
+fi
 
 #fix names
 mv GRAPES NAPA || exit
